Use literal values for macro bindings
C preprocessor macros have no inherent type. Most are represented in
Swift as `Int32` as an artifact of the Swift-C bridging layer, but are
generally used throughout the API as `UInt32` or other numeric types.

Hard-coding literal values with the correct Swift types eliminates
potentially unsafe type casting.

+++ Tests/swift-libgit2Tests/Index-Tests.swift
@@ -851,7 +851,7 @@ final class IndexTests: XCTestCaseStopOnFail
     
     func testGitIndexEntryNameMask() throws
     {
-        XCTAssertEqual(gitIndexEntryNameMask, GIT_INDEX_ENTRY_NAMEMASK)
+        XCTAssertEqual(gitIndexEntryNameMask, UInt32(GIT_INDEX_ENTRY_NAMEMASK))
     }
     
     
@@ -893,7 +893,7 @@ final class IndexTests: XCTestCaseStopOnFail
     
     func testGitIndexEntryStageMask() throws
     {
-        XCTAssertEqual(gitIndexEntryStageMask, GIT_INDEX_ENTRY_STAGEMASK)
+        XCTAssertEqual(gitIndexEntryStageMask, UInt32(GIT_INDEX_ENTRY_STAGEMASK))
     }
